Release version 28.3 2022-01-03
The input is no longer written to stdout.

By default, the Rule 2b size is written to stdout.

With -v, Rule 2b size, Rule 2a size and matched keyword count are written to stdout.

A non-ASCII or a NUL byte will issue a warning instead of exiting.

Warnings are written to stderr.

The source may come from stdin or a file argument.

New Exit codes:

    0 - soruce code is within Rule 2a and Rule 2b limits
    1 - soruce code larger than Rule 2a and/or Rule 2b limits
    2 - -h used and help printed
    3 - -V used and version printed
    >= 4 - some internal error occurred