Release version 0.26 2022-01-17
Fixed problems with error messages related to tar, cp and ls.

If the tar, cp and/or ls utility is not in the historic location,
look for it in an alternate location (for those distributions where
/usr/bin != /bin AND they have the mistake of moving the utility).

Improved the languare related to the tar -J flag.

Improved and/or fixed the diagnostic language related
to tar, cp and ls utilities.