Fix some gcc warnings
Without declaring a standard, %lld is not supported.
We cast to long mostly because today most longs will
support filesizes.  Even where they don't such cases
are outside rule Rule 2a and/or Rule 2b and only warning
or error messages are impacted.

Removed struct timezone tzp as the tzp was only
causing compabilty problems and was not even used.