Make it easier to update existing entries (#29)
Add -o <output> option that writes the answers to the specified file (if
it does not already exist; it is not overwritten). Then one can do e.g.
`cat answers.txt | ./mkiocccentry ...' so that you don't have to
repeatedly type the same information like id, name, abstract etc.

There is probably room for clean-up here maybe particularly the calls to
warn/fpara/para and possibly modularising the writing to the output file
but this was done quickly for the idea (I make a lot of updates during
the contest so having this feature would be highly useful).