Typo fixes + remove useless use of cat for answers
There were some typos: ansewrs > answers.

When I added the answers file code I said to do `cat answers |
./mkiocccentry ...' but I should have (and knew better) specified to
just do `./mkiocccentry ...  < answers' and I've updated the strings in
the code to do this.

I've also added my name to the right of my handle (I have also added my
name to my profile so that hopefully this will not be a problem in the
future with commits). Thank you Landon for adding my name to the
comments in the first place - I really appreciate it!