Remarks in tarball fix (#32)
* Typo fixes + remove useless use of cat for answers

There were some typos: ansewrs > answers.

When I added the answers file code I said to do `cat answers |
./mkiocccentry ...' but I should have (and knew better) specified to
just do `./mkiocccentry ...  < answers' and I've updated the strings in
the code to do this.

I've also added my name to the right of my handle (I have also added my
name to my profile so that hopefully this will not be a problem in the
future with commits). Thank you Landon for adding my name to the
comments in the first place - I really appreciate it!

* Open answers file after making entry directory

The problem with opening it first is that if the entry directory already
exists and we have created the answers file then the answers file will
have invalid data. Thus we open it after and then write the previously
obtained information to the file.

* Fix file name remarks_md to be remarks.md

I can only assume that this was a typo.