Ignoring warnings with answer file fixes
If answer file specified and user ignores warnings write the
confirmation to the answers file.

When ignoring warnings and specifying the answers file we print a
warning that if they change this (fix the warnings or cause additional
warnings) that the results are likely to be incorrect and strongly
recommend that they re-enter the answers.

When showing how to use the answers file to update an entry warn them
that if they had ignored warnings and fixed them (or if new warnings are
triggered) then they are strongly recommended to re-enter the answers.