Add direct support to read answers file
If you specify -i answers it'll try reading from the file answers.
Cannot specify -i and -a at the same time. Uses freopen(). I read in the
git log that there might be a problem with that function on some systems
but I'm not familiar with the specifics so if this is a desirable
feature (it seems that having it built-in would make it easier to use
for most people but maybe not) that might have to be addressed.