Ignoring warnings with answer file fixes (#34)
* Ignoring warnings with answer file fixes

If answer file specified and user ignores warnings write the
confirmation to the answers file.

When ignoring warnings and specifying the answers file we print a
warning that if they change this (fix the warnings or cause additional
warnings) that the results are likely to be incorrect and strongly
recommend that they re-enter the answers.

When showing how to use the answers file to update an entry warn them
that if they had ignored warnings and fixed them (or if new warnings are
triggered) then they are strongly recommended to re-enter the answers.

* Add direct support to read answers file

If you specify -i answers it'll try reading from the file answers.
Cannot specify -i and -a at the same time. Uses freopen(). I read in the
git log that there might be a problem with that function on some systems
but I'm not familiar with the specifics so if this is a desirable
feature (it seems that having it built-in would make it easier to use
for most people but maybe not) that might have to be addressed.

* Always call fclose() on answers file

We cannot write the final answers to the answers file if writing answers
is not specified but we still should close it if we're only reading
answers.

* Comment update

Only to make it so git sees a change so I can fix the bug where if -i
was specified it wouldn't close answers file.