Bad file descriptor fix + warn on answers I/O
If -i is specified you can get a bad file descriptor when confirming
that the tarball is correct. I _thought_ I fixed this yesterday but
there's another case: because I use freopen() to make stdin refer to the
answers file the answers file has to be open until _after_ everything
has been read (the confirmation of the tarball being correct comes from
stdin or answers file so we can't close answers file until after this is
done). Thus form the tarball before finalising the answers file.

I also added a counter of I/O errors on the answers file.

Additionally I have added a warning for answers file even if no warnings
were ignored, telling the user that if they do happen to add or change a
file that does trigger a warning that it will cause incorrect results.