Answers file enhancements & fixes wrt warnings
When using the '-a answers' option if there are warnings e.g. empty
prog.c the warnings are displayed prior to proceeding (e.g. the warning
is displayed in check_prog_c()) as if -a was not used. However when
using e.g. '-i answers' the warnings (IF there are warnings, new or not)
the warnings are NOT shown in the check functions but rather at the end
right before forming the tarball. The warnings are shown one after the
other, with the full text, so that the user understands the reasoning
behind it and why it's significant (prior to this commit the full
warning was displayed but it was automatically ignored, scrolling by,
which means the user probably wouldn't see it and then there was only a
short warning).

The warning messages have all been modularised: a function for each
warning (the rule 2a warnings are in the same function with a mode
parameter passed in to say which warning to refer to). This was
necessary to not have duplication of the warnings and their text.

Report I/O errors on the answers file when using -a rather than -i.
Perhaps there should be warnings on this with -i too e.g. if the answers
file has a read error in one or more of the answers?

Remove echo of answers since the answers file no longer has the ignoring
of warnings (thus the user actually types the 'y' so it is shown).

There might be other changes in this commit that I'm not remembering (I
wrote these changes against Ilya's patch the other day).