Upper case 'strongly' in warning messages
I feel this is useful as these are important and this should make the
significance stand out just a bit more.

In my previous commit I said there might have been something else I
forgot and I remember one of those things now: even with the answers
file always prompt if the tarball looks correct; it seems to me that
despite the fact the answers file automates the input of answers the
tool should always prompt if the tarball is correct: it would be too
easy to assume everything is okay because the answers file didn't result
in any problems; the fact new files can be added (or some files removed)
means the tarball can have different files so it should be checked that
it's correct before proceeding.