Fix mkiocccentry-test.sh
Now works after commit c9b31d416c82932daa0e6ec4dd8d094d2d3edb25. Due to
the version tag and EOF marker redirecting stdin (or cat answers |
./mkiocccentry ...') no longer works so use -i in the final test.

Due to the fact that the program always prompts if the tarball is
correct (even with answers file) pipe yes to the second and third
invocations of mkiocccentry: this way the program receives 'y' in reply
to the confirmations of (1) an empty prog.c; and (2) that the tarball is
correct.