Minor typo fixes
The one that might be worth pointing out is that in
check_extra_data_files() remarks.md was referred to instead of the
actual file name. Since the code uses the fpara() function and I'm not
too familiar with its internals I have it just say e.g. 'The file, while
it exists, ...' rather than refer to the file itself (the err() function
does actually name it). This does bring up a funny error message that
maybe should be fixed: 'The file, while it exists, is not a file.' but
I'll leave that to another person as it amuses me greatly :) (maybe I'll
fix it if nobody else does but at another time).