Minor fixes (#46)
* Answers file enhancements & fixes wrt warnings

When using the '-a answers' option if there are warnings e.g. empty
prog.c the warnings are displayed prior to proceeding (e.g. the warning
is displayed in check_prog_c()) as if -a was not used. However when
using e.g. '-i answers' the warnings (IF there are warnings, new or not)
the warnings are NOT shown in the check functions but rather at the end
right before forming the tarball. The warnings are shown one after the
other, with the full text, so that the user understands the reasoning
behind it and why it's significant (prior to this commit the full
warning was displayed but it was automatically ignored, scrolling by,
which means the user probably wouldn't see it and then there was only a
short warning).

The warning messages have all been modularised: a function for each
warning (the rule 2a warnings are in the same function with a mode
parameter passed in to say which warning to refer to). This was
necessary to not have duplication of the warnings and their text.

Report I/O errors on the answers file when using -a rather than -i.
Perhaps there should be warnings on this with -i too e.g. if the answers
file has a read error in one or more of the answers?

Remove echo of answers since the answers file no longer has the ignoring
of warnings (thus the user actually types the 'y' so it is shown).

There might be other changes in this commit that I'm not remembering (I
wrote these changes against Ilya's patch the other day).

* Upper case 'strongly' in warning messages

I feel this is useful as these are important and this should make the
significance stand out just a bit more.

In my previous commit I said there might have been something else I
forgot and I remember one of those things now: even with the answers
file always prompt if the tarball looks correct; it seems to me that
despite the fact the answers file automates the input of answers the
tool should always prompt if the tarball is correct: it would be too
easy to assume everything is okay because the answers file didn't result
in any problems; the fact new files can be added (or some files removed)
means the tarball can have different files so it should be checked that
it's correct before proceeding.

* Fix mkiocccentry-test.sh

Now works after commit c9b31d416c82932daa0e6ec4dd8d094d2d3edb25. Due to
the version tag and EOF marker redirecting stdin (or cat answers |
./mkiocccentry ...') no longer works so use -i in the final test.

Due to the fact that the program always prompts if the tarball is
correct (even with answers file) pipe yes to the second and third
invocations of mkiocccentry: this way the program receives 'y' in reply
to the confirmations of (1) an empty prog.c; and (2) that the tarball is
correct.

* Add -D_XOPEN_SOURCE to CFLAGS

When this feature test macro is defined before including stdlib.h it
exposes the WEXITSTATUS() etc. macros which are used in the code wrt
system(3) calls.

I also fixed a typo in the comments in the Makefile.

Co-authored-by: Landon Curt Noll <lcn2@users.noreply.github.com>