Make answers file version and EOF marker #defines
This is so that it's easier to update the version. The version format
could be changed but I made it the following:

	"MKIOCCCENTRY_ANSWERS-YYYY-##.##"

...where the final two numbers are major.minor, currently at 00.00 since
real programmers start with 0! :)

The EOF marker now is "ANSWERS_EOF" rather than "ANSWERS_END".

This required a change in the mkiocccentry-test.sh script for obvious
reasons. Since the version is no longer hard coded (well - it's now a
define) the script has to retrieve the version from the mkiocccentry.c
file. Currently it's done in the answers() function like:

	echo $(grep -E '^#define MKIOCCCENTRY_ANSWERS_VER' mkiocccentry.c|awk '{print $3  ;}'|sed 's/"//g')

...right before the cat << "EOF". Perhaps this can be improved but it
works assuming awk and sed are installed (which I should hope are).
Maybe someone knows of a way to do this in bash itself and can make it
better but I don't know if this is possible.