Makefile warning fixes .info.json file updates
Show the complete warnings for invalid Makefiles (i.e. missing targets)
if -i answers is specified (before it was just the final message about
Makefile not being valid).

Write to the .info.json file the rules found in the Makefile (or not
found as the case might be): this way the judges can see which warnings
were ignored/what targets weren't specified rather than just the user
ignored warning(s).

At the same time the tarball path (basename) is now written to the
.info.json file. I'm not sure if this is strictly necessary but the
thought crossed my mind that this might be of use in some way or
another.

I think that's all this commit includes.