Fixes and enhancements (#48)
* Make answers file version and EOF marker #defines

This is so that it's easier to update the version. The version format
could be changed but I made it the following:

	"MKIOCCCENTRY_ANSWERS-YYYY-##.##"

...where the final two numbers are major.minor, currently at 00.00 since
real programmers start with 0! :)

The EOF marker now is "ANSWERS_EOF" rather than "ANSWERS_END".

This required a change in the mkiocccentry-test.sh script for obvious
reasons. Since the version is no longer hard coded (well - it's now a
define) the script has to retrieve the version from the mkiocccentry.c
file. Currently it's done in the answers() function like:

	echo $(grep -E '^#define MKIOCCCENTRY_ANSWERS_VER' mkiocccentry.c|awk '{print $3  ;}'|sed 's/"//g')

...right before the cat << "EOF". Perhaps this can be improved but it
works assuming awk and sed are installed (which I should hope are).
Maybe someone knows of a way to do this in bash itself and can make it
better but I don't know if this is possible.

* Fix compilation under CentOS

Under CentOS with the _XOPEN_SOURCE feature test macro (to expose
WEXITSTATUS()) enabled ssize_t is no longer available. Thus #include
sys/types.h and sys/wait.h instead. Tested under CentOS and macOS.

* Makefile warning fixes .info.json file updates

Show the complete warnings for invalid Makefiles (i.e. missing targets)
if -i answers is specified (before it was just the final message about
Makefile not being valid).

Write to the .info.json file the rules found in the Makefile (or not
found as the case might be): this way the judges can see which warnings
were ignored/what targets weren't specified rather than just the user
ignored warning(s).

At the same time the tarball path (basename) is now written to the
.info.json file. I'm not sure if this is strictly necessary but the
thought crossed my mind that this might be of use in some way or
another.

I think that's all this commit includes.

* Simplify retrieval of answers version in test

No need to use awk; just use cut instead.