Resolve issue #47 + add ignore warnings option
If -a is specified and -A is NOT specified, warn the user if the answers
file already exists and ask them if they want to overwrite it. If they
say yes the answers file will be overwritten. If -A is specified or the
answers file does not already exist, the answers file is written to
whether or not it exists. -A implies -a (falls through in switch).

To make it easier to test add an option that allows the user to ignore
warnings. When this option -W is specified, ignore warnings once and
warn them that they're doing this, telling them how to fix this (in case
they did it by accident) and reminding them that the judges will not
ignore the warnings.