Clean up mkiocccentry-test.sh + add new test
Quote variables e.g. "${work_dir}" instead of $work_dir.

Move the retrieval of the mkiocccentry version out of the answers()
function and put it right before the call to answers(). This makes
shellcheck happy: otherwise it reports useless echo but in fact it's NOT
useless; removing echo from the function would not add the version to
the answers file.

Add another test case: an empty prog.c but ignore the warning about it
with the -W option in the previous commit. Still uses `yes | ...`
because the program would otherwise prompt the user that the tarball
looks correct.