Add utils.{h,c} with some of the helper functions
The following functions are now in utils.h/utils.c:

    bool exists(char const *path);
    bool is_file(char const *path);
    bool is_exec(char const *path);
    bool is_dir(char const *path);
    bool is_read(char const *path);
    bool is_write(char const *path);
    ssize_t file_size(char const *path);
    ssize_t readline(char **linep, FILE * stream);
    char *readline_dup(char **linep, bool strip, size_t *lenp, FILE * stream);
    void para(char const *line, ...);
    void fpara(FILE * stream, char const *line, ...);
    char *prompt(char const *str, size_t *lenp);
    ssize_t file_size(char const *path);

Additionally we obtain the 'bool' in this file and 'FILE *input_stream'
is also in utils.c/utils.h.

The purpose of this change is to help modularise these functions as it
appears that at least for the tarball checker it might be needed to use
C (since there doesn't seem to be a portable way to get the file size
for example and since the function is already written it might be good
to have it in a separate file).

I have only a very basic header in the files as I'm not sure what
comments to include/what to say (there's some humour there though). I
leave this up to the judges (which I guess means Landon).