Resolve issue #47 + add ignore warnings option (#50)
* Resolve issue #47 + add ignore warnings option

If -a is specified and -A is NOT specified, warn the user if the answers
file already exists and ask them if they want to overwrite it. If they
say yes the answers file will be overwritten. If -A is specified or the
answers file does not already exist, the answers file is written to
whether or not it exists. -A implies -a (falls through in switch).

To make it easier to test add an option that allows the user to ignore
warnings. When this option -W is specified, ignore warnings once and
warn them that they're doing this, telling them how to fix this (in case
they did it by accident) and reminding them that the judges will not
ignore the warnings.

* Clean up mkiocccentry-test.sh + add new test

Quote variables e.g. "${work_dir}" instead of $work_dir.

Move the retrieval of the mkiocccentry version out of the answers()
function and put it right before the call to answers(). This makes
shellcheck happy: otherwise it reports useless echo but in fact it's NOT
useless; removing echo from the function would not add the version to
the answers file.

Add another test case: an empty prog.c but ignore the warning about it
with the -W option in the previous commit. Still uses `yes | ...`
because the program would otherwise prompt the user that the tarball
looks correct.

* Fix clean up of mkiocccentry-test.sh

There were some issues with the clean up that I missed, ironically
making shellcheck unhappy again. This commit should resolve the
problems.