Add make depend (#54)
This shows how it works. I have left out the rule_count.o because the
rule is special. Anyway when a new object file is added you can add it
to the OBJFILES variable and then run make depend. It uses awk but it
could be done with cut too (and probably should be fixed). I'm sure you
get the idea behind this and you can adapt it to your purposes if
necessary.