Move cmdprintf() to utils (#55)
* Add make depend

This shows how it works. I have left out the rule_count.o because the
rule is special. Anyway when a new object file is added you can add it
to the OBJFILES variable and then run make depend. It uses awk but it
could be done with cut too (and probably should be fixed). I'm sure you
get the idea behind this and you can adapt it to your purposes if
necessary.

* Move cmdprintf() to util.c/util.h

For future use in the tarball checker utility which is supposed to call
the filenamechk tool.

* Typo fixes in mkiocccentry.1