Initial txzchk (issue #42)
This is incomplete but the following changes were made for this commit.
It's a lot and if there are any problems (as is possible with such
changes) I will fix them. I've tested various tarballs (size, not
starting with a directory, having any special files etc.) but much could
be done still. Because there's a lot I'm going to make a comment in the
GitHub pull request. I'm sorry if for all the text there but I think
it's important! I'll fix any problems and I'll tackle the things not
completed another time (probably not today) if the judges (or someone
else) does not (there are a number of things left to do).

* mkiocccentry

    Change usage message a bit: since the line was so long for the
    synopsis I just have [options] and below the options are shown (as
    they always were).

    The tool now runs (in the form_tarball() function) the txzchk tool
    (see below).

    The locating of the tools (tar, cp, ls) were moved to util.c/util.h
    in a separate function (that also finds other tools).

* mkiocccentry.1

    Formatting fixes.

* txzchk

For discussion here see the comment in the pull request. This will
include possible issues as well as things that aren't yet implemented.

* Makefile

    Add missing util.o to make clean rule.

    Add txzchk to Makefile.

Additionally there were some typo fixes.