More issues resolved in txzchk
Detect setuid/setgid/sticky bits in entries in tarballs. This is the way
that popped into my head last night so maybe it could be better or maybe
I've missed something but I think it might well suffice.

Fix issue with tarball non-numerical UID/GID
entry.test.10.9876543210.txz.

Make the -f fnamchk option -F to be in sync with the mkiocccentry tool
and fix the boolean saying that this option was specified to start out
with false (was true).

Test that we can execute fnamchk.

Run fnamchk _after_ testing the tarball contents. This is so that (0)
the judges can see the problems in the tarball as well as if the name is
valid (if fnamchk ends up saying it's invalid it exits so the txzchk
tool doesn't continue); and (1) so it's easier to test this tool. Make
sure to report the number of issues prior to running the fnamchk as
well.

Don't falsely report that all tests passed when they didn't (in main()).

The way the fields are extracted are via strtok() but perhaps a better
way could be formed and perhaps for later tests this will be necessary
(I don't know - just pointing out how it's done currently). I use
strdup() on the original line to show the line in question in warnings
(since strtok() modifies the original string).

Fix bug where the fnamchk tool could not always be found (when invoked
from mkiocccentry) since that tool switches directories. It appears that
this was partly done in the previous commit. Note that this has to be
tested more thoroughly! I somehow suspect that this can be broken but at
least under the normal conditions (running mkiocccentry from the
directory just above the work directory) it seems to function fine.