bool test true/false changes + ?_flag_used renamed
Instead of doing == true and == false just have e.g. if (test) and
if (!test).

To be consistent with txzchk_flag_used make the other flag_used
variables in this form too. This can be of help when looking later at
the code as it can remind exactly what the flag means.