Make txz filename format consistent w/dir name
In anticipation of the txzchk verifying that the directory name is
correct, change the format of the filenames to be:

    "entry." ("test" or a valid UUID) "\-" entry_number "." timestamp ".txz"

This is because the directory name in the tarball is "test" or valid UUID
followed by a '-' and then the entry number instead of a '.'. The tool
fnamchk used to print out either "test" or a valid UUID followed by a
'.' followed by the entry number; now it uses a '-' instead of a '.'
which is exactly how the directory name is. Thus the txzchk tool can get
the expected directory name in the tarball by calling the fnamchk tool.

What's not done yet is moving fnamchk call to before the test for the
directory name tests. If however I extract all the fields in advance I
can perform all the tests that need to be done (except the directory
name) before calling fnamchk and in this way I can still have fnamchk
return an error if the filename is not correct. On the other hand
fnamchk could be changed (this was not done) to warn on invalid
filenames rather than error out (thus preventing txzchk from continuing
which is why fnamchk call was moved to after all the tests txzchk
performs).

Both fnamchk.1 and mkiocccentry.c have been updated to account for this
change.