File sizes test added to txzchk
txzchk now tallies the size of all the files in the tarball and rounds
it up to the next multiple of 1024 and warns if it's > MAX_DIR_KSIZE
(also if it's < 0 - which is an error).

The function that rounds it up (round_to_multiple()) uses off_t as it's
meant for file sizes but perhaps it could be changed to be more general
(maybe long long).