txzchk updates and fixes (including several checks added) (#59)
* bool test true/false changes + ?_flag_used renamed

Instead of doing == true and == false just have e.g. if (test) and
if (!test).

To be consistent with txzchk_flag_used make the other flag_used
variables in this form too. This can be of help when looking later at
the code as it can remind exactly what the flag means.

* Make txz filename format consistent w/dir name

In anticipation of the txzchk verifying that the directory name is
correct, change the format of the filenames to be:

    "entry." ("test" or a valid UUID) "\-" entry_number "." timestamp ".txz"

This is because the directory name in the tarball is "test" or valid UUID
followed by a '-' and then the entry number instead of a '.'. The tool
fnamchk used to print out either "test" or a valid UUID followed by a
'.' followed by the entry number; now it uses a '-' instead of a '.'
which is exactly how the directory name is. Thus the txzchk tool can get
the expected directory name in the tarball by calling the fnamchk tool.

What's not done yet is moving fnamchk call to before the test for the
directory name tests. If however I extract all the fields in advance I
can perform all the tests that need to be done (except the directory
name) before calling fnamchk and in this way I can still have fnamchk
return an error if the filename is not correct. On the other hand
fnamchk could be changed (this was not done) to warn on invalid
filenames rather than error out (thus preventing txzchk from continuing
which is why fnamchk call was moved to after all the tests txzchk
performs).

Both fnamchk.1 and mkiocccentry.c have been updated to account for this
change.

* t_flag_used -> tar_flag_used (txzchk.c)

Previously only fnamchk bool had this format.

* Bump version in fnamchk and mkiocccentry

The change in filename format seems like a reason to update the version.

* File sizes test added to txzchk

txzchk now tallies the size of all the files in the tarball and rounds
it up to the next multiple of 1024 and warns if it's > MAX_DIR_KSIZE
(also if it's < 0 - which is an error).

The function that rounds it up (round_to_multiple()) uses off_t as it's
meant for file sizes but perhaps it could be changed to be more general
(maybe long long).

* Bump version of txzchk (for recent test added)

* Add several checks to txzchk

The following tarballs and their issues are now detected. Please note
that since the tarball filename format has changed slightly the filename
will be different from what's actually downloadable on GitHub (I have
attempted to rename them below to fix).

    entry.test-1.9876543210.txz: duplicate filenames detected.

    entry.test-4.9876543210.txz: detect dot files that aren't
    '.author.json' and '.info.json'.

    entry.test-9.9876543210.txz: detect empty critical files.

* Name the invalid dot files in the warning

* Two more tests added to txzchk

Now detects filenames with '../' in the path.

It should also detect absolute paths but this is hard to test because
the tarball supplied had the initial '/' stripped and I had the same
problem when creating one.