Clean up and bug fix txzchk
Make warning messages more consistent in format: always show the tool's
name and the txzpath. For example if there's no .info.json file it'll
now show e.g.:

	Warning: txzchk: test.txz: no .info.json found

...whereas before it showed the function name and did not show the
tarball path (in this case the fnamchk would issue an error). I feel
that showing the utility name is more useful than 'check_tarball()'.
It's possible I have missed a warning. I tried to do the same with
errors but I spent less time working on those as those are fatal and the
program ends.

Removed the check_files() function; the tests are now all in
check_tarball().

check_tarball() now issues a warning if fnamchk fails so that we can
continue if it does fail (though currently it's the last step of the
function).

txzchk now always shows how many warnings were issued even if fnamchk
fails (which is considered an issue).

txzchk now exits with 1 if there are issues found.