More checks added to txzchk
fnamchk is now called before the other tests: this is so that we can (if
no error) retrieve the directory name for tests later on (when parsing
the output of tar). If fnamchk returns an error it will be reported but
txzchk won't be ended (if system(3) returns < 0 or 127 it will end):
instead the other tests will be run.

This commit detects some of the cases where the directory name was
wrong: if the entry number is 0 then a directory of test-3 will fail for
example. There might be some other tests that this commit adds.
Subdirectories are still undetected directly (if it triggers another
warning e.g. test-0/../../dir/.file will fail on both ../ and .file but
the fact dir/ is there will not yet be reported).