Add test for no directory in tarball
In other words a tarball with the contents of:

    -rw-r--r-T  0 cody   staff     217 Feb  6 09:49 ptr.c

...will now detect that ptr.c has no directory (in addition to the
special bit, the non-numerical UID and GID as well as the fact the
directory isn't correct and fnamchk would fail). The way this is done is
by checking for a '/' in the path. I'm not sure if this check tests
every possible scenario or if there's a way to make it so that the test
wrongly reports a problem.