txzchk: Clean up, bug fixes and more tests added (#60)
* Remove redundant tarball listing output

Since the txzchk is always executed by mkiocccentry and since txzchk
prints out the contents of the tarball (since it has to parse the output
of tar) there's no need to explicitly invoke tar -tJvf. If the txzchk
tool is changed so it doesn't show the output of the tar listing then
this will have to be added back but this way it's no longer showing the
same information twice.

Typo fix(es).

* Clean up and bug fix txzchk

Make warning messages more consistent in format: always show the tool's
name and the txzpath. For example if there's no .info.json file it'll
now show e.g.:

	Warning: txzchk: test.txz: no .info.json found

...whereas before it showed the function name and did not show the
tarball path (in this case the fnamchk would issue an error). I feel
that showing the utility name is more useful than 'check_tarball()'.
It's possible I have missed a warning. I tried to do the same with
errors but I spent less time working on those as those are fatal and the
program ends.

Removed the check_files() function; the tests are now all in
check_tarball().

check_tarball() now issues a warning if fnamchk fails so that we can
continue if it does fail (though currently it's the last step of the
function).

txzchk now always shows how many warnings were issued even if fnamchk
fails (which is considered an issue).

txzchk now exits with 1 if there are issues found.

* txzchk: Check for NUL bytes before end of line

* txzchk: warning prefix consistency

* More checks added to txzchk

fnamchk is now called before the other tests: this is so that we can (if
no error) retrieve the directory name for tests later on (when parsing
the output of tar). If fnamchk returns an error it will be reported but
txzchk won't be ended (if system(3) returns < 0 or 127 it will end):
instead the other tests will be run.

This commit detects some of the cases where the directory name was
wrong: if the entry number is 0 then a directory of test-3 will fail for
example. There might be some other tests that this commit adds.
Subdirectories are still undetected directly (if it triggers another
warning e.g. test-0/../../dir/.file will fail on both ../ and .file but
the fact dir/ is there will not yet be reported).

* Typo fix

* Remove unnecessary free

When moving the invocation of fnamchk to above the invocation of tar I
forgot to remove the free(cmd) so it was there twice (this was safe
because I set cmd = NULL). This has been removed.

* Add test for no directory in tarball

In other words a tarball with the contents of:

    -rw-r--r-T  0 cody   staff     217 Feb  6 09:49 ptr.c

...will now detect that ptr.c has no directory (in addition to the
special bit, the non-numerical UID and GID as well as the fact the
directory isn't correct and fnamchk would fail). The way this is done is
by checking for a '/' in the path. I'm not sure if this check tests
every possible scenario or if there's a way to make it so that the test
wrongly reports a problem.

* txzchk: Typo fix in comment