Always attempt to read to EOF in tarball listing
In other words if a NULL pointer is encountered (or empty line) in the
parsing of the line, continue to the next line (until EOF). If there's a
memory error we do abort. Although this should not happen in most cases
(once the tar listing output of linux is parsed properly) - since
usually tar will be used - it can be helpful for the -T option in
particular for testing different convoluted outputs etc. (in case a user
also tries to construct a tarball with problems and submits it).