Move parsing of line in tar to new function
In anticipation of fixing the parsing error under linux (due to
different tar output) I have moved the parsing line code to a separate
function. It still does not parse linux output correctly but that's what
I'll work on next. Both make test and running the tool on the file make
test runs succeed; running the tool on the tarballs with problems detect
the issues that are currently finished.