txzchk: portability, bug fixes and enhancements (#61)
* Portability fixes wrt off_t and long long

Compile using C99 for support of long long under CentOS. This change
caused other problems under CentOS and fixing those caused problems
under macOS. Enabling certain feature test macros solves the problem for
CentOS and using setvbuf() instead of setlinebuf() solved another
problem. We also have to #include <time.h> in limit_ioccc.h for this
change as the other changes cause a problem under macOS.

Because off_t is different sizes in different systems (macOS it's long
long; CentOS it's not) now cast to unsigned long when printing (we test
for < 0 prior to printing so this should be okay but this can always be
changed).

All the files should now compile cleanly under linux and macOS after
doing a make clobber.

* Change bool tests in util.c

To be consistent with the other files from commit
6712f8f9fb7afff67030b6694a3eb53268f04ccf do if (value) and if (!value)
instead of == true and == false. That should be the last file that needs
to be updated. Note that there's a comment that has == true but this is
okay and makes sense.

* make install now installs man pages

* Add -T (text file mode) option to txzchk

The tool now allows assuming that the txzpath is a text file, mostly for
testing purposes (different platforms have different tar list formats
for example).

* Always attempt to read to EOF in tarball listing

In other words if a NULL pointer is encountered (or empty line) in the
parsing of the line, continue to the next line (until EOF). If there's a
memory error we do abort. Although this should not happen in most cases
(once the tar listing output of linux is parsed properly) - since
usually tar will be used - it can be helpful for the -T option in
particular for testing different convoluted outputs etc. (in case a user
also tries to construct a tarball with problems and submits it).

* Move parsing of line in tar to new function

In anticipation of fixing the parsing error under linux (due to
different tar output) I have moved the parsing line code to a separate
function. It still does not parse linux output correctly but that's what
I'll work on next. Both make test and running the tool on the file make
test runs succeed; running the tool on the tarballs with problems detect
the issues that are currently finished.